Refactor FastAPI schemas into dedicated modules
