+++ backend/schemas/__init__.py
@@ -1,0 +1,65 @@
+from .ai_system import AISystem
+from .ai_system_update import AISystemUpdate
+from .audit import Audit
+from .contact import Contact
+from .contact_method import ContactMethod
+from .contact_preference import ContactPreference
+from .dashboard_overview import DashboardOverview
+from .deliverable import Deliverable
+from .deliverable_assignment import DeliverableAssignment
+from .deliverable_template import DeliverableTemplate
+from .deliverable_update import DeliverableUpdate
+from .evidence import Evidence
+from .incident import Incident
+from .incident_update import IncidentUpdate
+from .login_payload import LoginPayload
+from .login_result import LoginResult
+from .org_structure import OrgStructure
+from .pending_activity import PendingActivity
+from .raci_entry import RACIEntry
+from .risk_assessment import RiskAssessment
+from .risk_wizard_config import RiskWizardConfig
+from .settings import Settings
+from .sign_in_payload import SignInPayload
+from .sign_in_response import SignInResponse
+from .sso_login_payload import SSOLoginPayload
+from .task import Task
+from .task_update import TaskUpdate
+from .team_member import TeamMember
+from .technical_dossier import TechnicalDossier
+from .technical_dossier_template import TechnicalDossierTemplate
+from .user import User
+
+__all__ = [
+    "AISystem",
+    "AISystemUpdate",
+    "Audit",
+    "Contact",
+    "ContactMethod",
+    "ContactPreference",
+    "DashboardOverview",
+    "Deliverable",
+    "DeliverableAssignment",
+    "DeliverableTemplate",
+    "DeliverableUpdate",
+    "Evidence",
+    "Incident",
+    "IncidentUpdate",
+    "LoginPayload",
+    "LoginResult",
+    "OrgStructure",
+    "PendingActivity",
+    "RACIEntry",
+    "RiskAssessment",
+    "RiskWizardConfig",
+    "Settings",
+    "SignInPayload",
+    "SignInResponse",
+    "SSOLoginPayload",
+    "Task",
+    "TaskUpdate",
+    "TeamMember",
+    "TechnicalDossier",
+    "TechnicalDossierTemplate",
+    "User",
+]
