+++ backend/schemas/contact.py
@@ -1,0 +1,11 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class Contact(BaseModel):
+    id: str
+    project_id: str
+    name: str
+    role: str
+    email: Optional[str] = None
