+++ backend/schemas/contact_method.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class ContactMethod(str, Enum):
+    email = "email"
+    sms = "sms"
+    whatsapp = "whatsapp"
+    slack = "slack"
