+++ backend/schemas/deliverable_assignment.py
@@ -1,0 +1,8 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class DeliverableAssignment(BaseModel):
+    assignee: str
+    due_date: Optional[str] = None
