+++ backend/schemas/deliverable_template.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class DeliverableTemplate(BaseModel):
+    id: str
+    name: str
+    type: str
