+++ backend/schemas/incident_update.py
@@ -1,0 +1,10 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class IncidentUpdate(BaseModel):
+    severity: Optional[str] = None
+    status: Optional[str] = None
+    title: Optional[str] = None
+    description: Optional[str] = None
