+++ backend/schemas/login_payload.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class LoginPayload(BaseModel):
+    company: str
+    email: str
+    password: str
