+++ backend/schemas/login_result.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+from .user import User
+
+
+class LoginResult(BaseModel):
+    token: str
+    user: User
