+++ backend/schemas/pending_activity.py
@@ -1,0 +1,9 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class PendingActivity(BaseModel):
+    id: str
+    description: str
+    due_date: Optional[str] = None
