+++ backend/schemas/risk_assessment.py
@@ -1,0 +1,11 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class RiskAssessment(BaseModel):
+    id: str
+    system_id: str
+    date: str
+    classification: str
+    justification: Optional[str] = None
