+++ backend/schemas/sign_in_response.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+from .user import User
+
+
+class SignInResponse(BaseModel):
+    user: User
+    temporary_password: str
+    message: str
