+++ backend/schemas/sso_login_payload.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class SSOLoginPayload(BaseModel):
+    company: str
+    email: str
+    provider: str
