+++ backend/schemas/task.py
@@ -1,0 +1,12 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class Task(BaseModel):
+    id: str
+    system_id: str
+    title: str
+    status: str
+    assignee: Optional[str] = None
+    due_date: Optional[str] = None
