+++ backend/schemas/task_update.py
@@ -1,0 +1,10 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class TaskUpdate(BaseModel):
+    title: Optional[str] = None
+    status: Optional[str] = None
+    assignee: Optional[str] = None
+    due_date: Optional[str] = None
