+++ backend/schemas/team_member.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+
+class TeamMember(BaseModel):
+    id: str
+    system_id: str
+    name: str
+    role: str
