+++ backend/schemas/technical_dossier.py
@@ -1,0 +1,8 @@
+from typing import Dict
+
+from pydantic import BaseModel, Field
+
+
+class TechnicalDossier(BaseModel):
+    system_id: str
+    fields: Dict[str, str] = Field(default_factory=dict)
